backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1  # 👈 ЯВНО УКАЗЫВАЕМ ВЕРСИЮ bcrypt
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

#fastapi
#uvicorn[standard]
#sqlalchemy
#python-multipart
#pydantic
#python-jose[cryptography]==3.3.0   # Для JWT
#passlib[bcrypt]==1.7.4             # Для хеширования паролей
#python-multipart==0.0.6            # Для форм
#pydantic-settings==2.1.0           # Для настроек
#email-validator==2.1.0             # Для валидации email
